elements: from_state_vector don't fail in e=nan or e>1.0 cases.
This can happen in radial, parabolic, and hyperbolic cases, which we want to
test, so log a warning and continue with M0 = 0.

In the common case, assert that M0 is not nan.

Also fix up some nonsensical code and simplify with a bunch of asserts.

test_orbital: Now properly test the zero case.