test_orbital: Added test_from_state_vector_roundtrips.
This tests 20 cases of all sorts of r, v pairs to do a roundtrip, so we can test
a lot of cases without having to manually write out all of the elements for each
one. All r, v pairs should theoretically roundtrip via from_state_vector.

Many of the cases are failing (and commented out) due to bugs and missing
support for parabolic, hyperbolic and radial orbits.

Note that this is a white-box test where we are deliberately targeting the
various branches inside elements_from_state_vector, to test all the edge cases.
For example, this exercises the buggy code path also tested by the arg_pe_gt_180
tests.